+++ .github/workflows/deploy.yml
@@ -44,7 +44,3 @@ jobs:
           vercel-org-id: ${{ secrets.VERCEL_ORG_ID }}
           vercel-project-id: ${{ secrets.VERCEL_PROJECT_ID }}
           vercel-args: '--prod'
-          working-directory: ./
-        retry:
-          attempts: 3
-          delay: 10s
