+++ src/components/Footer.tsx
@@ -16,18 +16,18 @@ const Footer = () => {
               </div>
               <h2 className="text-xl font-bold text-white ml-2 font-display">PrepPair.Me</h2>
             </div>
-            <p className="text-gray-300 max-w-xs leading-relaxed">
+            <p className="text-muted-foreground max-w-xs">
               Your AI interview partner—turning resumes and job descriptions into tailored prep guides.
             </p>
             
             <div className="flex space-x-4 mt-4">
-              <a href="#" className="text-gray-300 hover:text-primary transition-colors">
+              <a href="#" className="text-muted-foreground hover:text-primary transition-colors">
                 <Twitter size={20} />
               </a>
-              <a href="#" className="text-gray-300 hover:text-primary transition-colors">
+              <a href="#" className="text-muted-foreground hover:text-primary transition-colors">
                 <Linkedin size={20} />
               </a>
-              <a href="#" className="text-gray-300 hover:text-primary transition-colors">
+              <a href="#" className="text-muted-foreground hover:text-primary transition-colors">
                 <Github size={20} />
               </a>
             </div>
@@ -35,41 +35,41 @@ const Footer = () => {
           
           <div className="grid grid-cols-2 md:grid-cols-3 gap-8">
             <div>
-              <h3 className="font-medium mb-4 text-white text-lg">Product</h3>
-              <ul className="space-y-3">
-                <li><Link href="/dashboard" className="text-gray-300 hover:text-primary font-medium">Dashboard</Link></li>
-                <li><Link href="/#features" className="text-gray-300 hover:text-primary font-medium">Features</Link></li>
-                <li><Link href="/pricing" className="text-gray-300 hover:text-primary font-medium">Pricing</Link></li>
+              <h3 className="font-medium mb-4 text-white">Product</h3>
+              <ul className="space-y-2">
+                <li><Link href="/dashboard" className="text-muted-foreground hover:text-primary">Dashboard</Link></li>
+                <li><Link href="/#features" className="text-muted-foreground hover:text-primary">Features</Link></li>
+                <li><Link href="/" className="text-muted-foreground hover:text-primary">Pricing</Link></li>
               </ul>
             </div>
             
             <div>
-              <h3 className="font-medium mb-4 text-white text-lg">Company</h3>
-              <ul className="space-y-3">
-                <li><Link href="/about" className="text-gray-300 hover:text-primary font-medium">About</Link></li>
-                <li><Link href="/blog" className="text-gray-300 hover:text-primary font-medium">Blog</Link></li>
-                <li><Link href="/contact" className="text-gray-300 hover:text-primary font-medium">Contact</Link></li>
+              <h3 className="font-medium mb-4 text-white">Company</h3>
+              <ul className="space-y-2">
+                <li><Link href="/about" className="text-muted-foreground hover:text-primary">About</Link></li>
+                <li><Link href="/" className="text-muted-foreground hover:text-primary">Blog</Link></li>
+                <li><Link href="/contact" className="text-muted-foreground hover:text-primary">Contact</Link></li>
               </ul>
             </div>
             
             <div>
-              <h3 className="font-medium mb-4 text-white text-lg">Legal</h3>
-              <ul className="space-y-3">
-                <li><Link href="/legal" className="text-gray-300 hover:text-primary font-medium">Privacy</Link></li>
-                <li><Link href="/legal" className="text-gray-300 hover:text-primary font-medium">Terms</Link></li>
-                <li><Link href="/legal" className="text-gray-300 hover:text-primary font-medium">Cookies</Link></li>
+              <h3 className="font-medium mb-4 text-white">Legal</h3>
+              <ul className="space-y-2">
+                <li><Link href="/legal" className="text-muted-foreground hover:text-primary">Privacy</Link></li>
+                <li><Link href="/legal" className="text-muted-foreground hover:text-primary">Terms</Link></li>
+                <li><Link href="/legal" className="text-muted-foreground hover:text-primary">Cookies</Link></li>
               </ul>
             </div>
           </div>
         </div>
         
         <div className="border-t border-white/10 mt-8 pt-8 flex flex-col md:flex-row justify-between items-center">
-          <p className="text-sm text-gray-300 font-medium">
+          <p className="text-sm text-muted-foreground">
             © {currentYear} PrepPair.Me. All rights reserved.
           </p>
           <div className="mt-4 md:mt-0 flex items-center">
             <MessageSquare className="h-4 w-4 text-primary mr-2" />
-            <span className="text-sm text-gray-300 font-medium">
+            <span className="text-sm text-muted-foreground">
               Built with ❤️ for job seekers everywhere
             </span>
           </div>
