--- src/components/calendar/components/EventForm.tsx
@@ -1,138 +1,0 @@
-
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from "@/components/ui/dialog";
-import { Label } from "@/components/ui/label";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { Loader2 } from "lucide-react";
-import { format } from "date-fns";
-
-interface EventFormProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onSubmit: (event: {
-    title: string;
-    description: string;
-    startTime: string;
-    endTime: string;
-    type: "interview" | "follow-up" | "reminder" | "other";
-  }) => Promise<void>;
-  isUpdating: boolean;
-}
-
-const EventForm = ({ isOpen, onClose, onSubmit, isUpdating }: EventFormProps) => {
-  const [newEvent, setNewEvent] = useState({
-    title: "",
-    description: "",
-    startTime: format(new Date(), "yyyy-MM-dd'T'HH:mm"),
-    endTime: format(new Date(new Date().getTime() + 60 * 60 * 1000), "yyyy-MM-dd'T'HH:mm"),
-    type: "reminder" as "interview" | "follow-up" | "reminder" | "other"
-  });
-
-  const handleSubmit = async () => {
-    await onSubmit(newEvent);
-    setNewEvent({
-      title: "",
-      description: "",
-      startTime: format(new Date(), "yyyy-MM-dd'T'HH:mm"),
-      endTime: format(new Date(new Date().getTime() + 60 * 60 * 1000), "yyyy-MM-dd'T'HH:mm"),
-      type: "reminder"
-    });
-  };
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Add New Event</DialogTitle>
-          <DialogDescription>
-            Create a new calendar event for your job search.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="grid gap-4 py-4">
-          <div className="space-y-2">
-            <Label htmlFor="title">Event Title</Label>
-            <Input
-              id="title"
-              placeholder="Event title"
-              value={newEvent.title}
-              onChange={e => setNewEvent(prev => ({ ...prev, title: e.target.value }))}
-              required
-            />
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="description">Description (Optional)</Label>
-            <Textarea
-              id="description"
-              placeholder="Event description"
-              value={newEvent.description}
-              onChange={e => setNewEvent(prev => ({ ...prev, description: e.target.value }))}
-            />
-          </div>
-          <div className="grid grid-cols-2 gap-4">
-            <div className="space-y-2">
-              <Label htmlFor="startTime">Start Time</Label>
-              <Input
-                id="startTime"
-                type="datetime-local"
-                value={newEvent.startTime}
-                onChange={e => setNewEvent(prev => ({ ...prev, startTime: e.target.value }))}
-                required
-              />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="endTime">End Time</Label>
-              <Input
-                id="endTime"
-                type="datetime-local"
-                value={newEvent.endTime}
-                onChange={e => setNewEvent(prev => ({ ...prev, endTime: e.target.value }))}
-                required
-              />
-            </div>
-          </div>
-          <div className="space-y-2">
-            <Label htmlFor="eventType">Event Type</Label>
-            <select
-              id="eventType"
-              className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-              value={newEvent.type}
-              onChange={e => {
-                const value = e.target.value;
-                if (
-                  value === "interview" ||
-                  value === "follow-up" ||
-                  value === "reminder" ||
-                  value === "other"
-                ) {
-                  setNewEvent(prev => ({ ...prev, type: value }));
-                }
-              }}
-            >
-              <option value="reminder">Reminder</option>
-              <option value="interview">Interview</option>
-              <option value="follow-up">Follow-up</option>
-              <option value="other">Other</option>
-            </select>
-          </div>
-        </div>
-        <DialogFooter>
-          <Button variant="outline" onClick={onClose}>
-            Cancel
-          </Button>
-          <Button 
-            onClick={handleSubmit} 
-            disabled={!newEvent.title || !newEvent.startTime || !newEvent.endTime || isUpdating}
-          >
-            {isUpdating ? <Loader2 className="mr-2 h-4 w-4 animate-spin" /> : null}
-            Add Event
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-export default EventForm;
-
