+++ src/integrations/supabase/types.ts
@@ -245,10 +245,6 @@ export type Database = {
         Args: Record<PropertyKey, never> | { user_id: string }
         Returns: boolean
       }
-      is_admin_user: {
-        Args: Record<PropertyKey, never>
-        Returns: boolean
-      }
     }
     Enums: {
       user_role: "admin" | "user"
