+++ supabase/config.toml
@@ -23,8 +23,8 @@ enabled = true
 
 [auth]
 enabled = true
-site_url = "https://preppair.me"
-additional_redirect_urls = ["https://preppair.me", "https://preppair.me/auth", "http://localhost:8080", "http://localhost:8080/auth"]
+site_url = "http://localhost:3000"
+additional_redirect_urls = ["https://localhost:3000"]
 jwt_expiry = 3600
 enable_refresh_token_rotation = true
 refresh_token_reuse_interval = 10
