+++ vercel.json
@@ -5,11 +5,13 @@
     { "handle": "filesystem" },
     { "src": "/(.*)", "dest": "/index.html" }
   ],
-  "buildCommand": "npm run build",
-  "outputDirectory": "dist",
+  "env": {
+    "VITE_OPENAI_API_KEY": "@openai_api_key",
+    "VITE_SUPABASE_URL": "@supabase_url",
+    "VITE_SUPABASE_ANON_KEY": "@supabase_anon_key"
+  },
   "github": {
     "silent": true,
     "autoAlias": true
-  },
-  "installCommand": "npm install --legacy-peer-deps"
+  }
 }
